rename: share hook moves to hooks folder
